Move all the auth components into a new feature directory
